Move vectorset API logic to a single module
